Demetri Day 6 and 7 (#14)
Looks goood. Passing in width to a point seems bad practice. You should hash based on a global or have a point be part of a grid system that can inject the width dependency. For advent, it's cool but I would be super scrutinizing of a dependency injection into each point. A point should exist outside of it's grid.